fix: given value must be an entity problem

+++ src/social-auth/github/github.strategy.ts
@@ -114,7 +114,7 @@ export class GithubStrategy implements ISocialAuthStrategy {
       'github.allow_override_authorization',
     );
 
-    if (allowOverrideAuthorization) {
+    if (exist && allowOverrideAuthorization) {
       await this.socialAuthRepository.remove(exist);
       exist = null;
     }
